feat: more strict type for count
1. Since we always operate only with a positive number of transfers
or the number of accounts, changing the account type to unsigned
saves us from a lot of checks and conversions.
2. Due to the fact that the linter works on the diff, and we make
some changes, this commit contains.

+++ pkg/engine/db/common.go
@@ -57,8 +57,8 @@ type commonCluster struct {
 
 	DBUrl string
 
-	connectionPoolSize int
-	addPool            int
+	connectionPoolSize uint64
+	addPool            uint64
 
 	sharded bool
 }
